requirements.txt
asgiref==3.8.1
Django==5.1
sqlparse==0.5.1
django-mptt==0.16.0
django-mptt-admin==2.7.0
